Consolidate common test code
