+++ dev/com.ibm.ws.microprofile.faulttolerance.cdi_fat/fat/src/com/ibm/websphere/microprofile/faulttolerance_fat/tests/stateless/bulkhead/BulkheadEJB.java
@@ -13,7 +13,7 @@ import javax.ejb.Stateless;
 
 import org.eclipse.microprofile.faulttolerance.Bulkhead;
 
-import com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.bulkhead.BarrierFactory.Barrier;
+import com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.BarrierFactory.Barrier;
 
 @Stateless
 public class BulkheadEJB {
