+++ dev/com.ibm.ws.microprofile.faulttolerance.cdi_fat/fat/src/com/ibm/websphere/microprofile/faulttolerance_fat/tests/stateless/bulkhead/BulkheadEJBTestServlet.java
@@ -9,16 +9,12 @@
  *******************************************************************************/
 package com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.bulkhead;
 
-import static org.hamcrest.Matchers.instanceOf;
-import static org.junit.Assert.assertThat;
+import static com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.ExecutionAssert.assertCompletes;
+import static com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.ExecutionAssert.assertThrowsEjbWrapped;
 
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
 import javax.annotation.Resource;
-import javax.ejb.EJBException;
 import javax.enterprise.concurrent.ManagedExecutorService;
 import javax.inject.Inject;
 import javax.servlet.annotation.WebServlet;
@@ -26,7 +22,8 @@ import javax.servlet.annotation.WebServlet;
 import org.eclipse.microprofile.faulttolerance.exceptions.BulkheadException;
 import org.junit.Test;
 
-import com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.bulkhead.BarrierFactory.Barrier;
+import com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.BarrierFactory;
+import com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.BarrierFactory.Barrier;
 
 import componenttest.annotation.ExpectedFFDC;
 import componenttest.app.FATServlet;
@@ -34,8 +31,6 @@ import componenttest.app.FATServlet;
 @SuppressWarnings("serial")
 @WebServlet("/bulkheadejb")
 public class BulkheadEJBTestServlet extends FATServlet {
-
-    private static final int COMPLETION_TIMEOUT = 1000;
 
     @Inject
     private BulkheadEJB bean;
@@ -80,38 +75,6 @@ public class BulkheadEJBTestServlet extends FATServlet {
             assertCompletes(f3);
             b5.complete();
             assertCompletes(f5);
-        }
-    }
-
-    private void assertThrowsEjbWrapped(Class<? extends Exception> expected, Future<?> future) {
-        EJBException e = assertThrows(EJBException.class, future);
-        assertThat("Wrapped exception has wrong type", e.getCause(), instanceOf(expected));
-    }
-
-    private <T extends Exception> T assertThrows(Class<T> expected, Future<?> future) {
-        try {
-            Object result = future.get(COMPLETION_TIMEOUT, TimeUnit.MILLISECONDS);
-            throw new AssertionError("Expected exception not thrown. Result: " + result);
-        } catch (ExecutionException e) {
-            Throwable t = e.getCause();
-            assertThat("Thrown exception has wrong type", t, instanceOf(expected));
-            return expected.cast(t);
-        } catch (InterruptedException e) {
-            throw new AssertionError("Interrupted waiting for future", e);
-        } catch (TimeoutException e) {
-            throw new AssertionError("Timeout waiting for future", e);
-        }
-    }
-
-    private <T> T assertCompletes(Future<T> future) {
-        try {
-            return future.get(COMPLETION_TIMEOUT, TimeUnit.MILLISECONDS);
-        } catch (ExecutionException e) {
-            throw new AssertionError("Future completed with exception: " + e, e);
-        } catch (InterruptedException e) {
-            throw new AssertionError("Interrupted waiting for future", e);
-        } catch (TimeoutException e) {
-            throw new AssertionError("Timeout waiting for future", e);
         }
     }
 }
